fix: extract MarkdownSegmentView and use it for subagent response events
Extract shared MarkdownSegmentView from ChatBubble's inline segment
rendering to eliminate duplication. Use it in SubagentDetailPanel for
rich markdown rendering (tables, code blocks, headings, lists) of
subagent response events. Make MarkdownTableView accept a maxWidth
parameter so tables expand to full width in the subagent panel.
